int -> double

+++ src/main/java/NullValueHandler.java
@@ -1,5 +1,6 @@
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
@@ -12,7 +13,7 @@ import org.apache.hadoop.mapreduce.lib.output.TextOutputFormat;
 import java.io.IOException;
 
 public class NullValueHandler {
-    public static class NullMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
+    public static class NullMapper extends Mapper<LongWritable, Text, Text, DoubleWritable> {
         @Override
         public void map(LongWritable key, Text values, Context context) throws IOException, InterruptedException {
             // value: user,movie,rating
@@ -20,12 +21,12 @@ public class NullValueHandler {
             if (user_move_rating.length<3){
                 return;
             }
-            int rating = Integer.parseInt(user_move_rating[2]);
-            context.write(new Text(user_move_rating[0]), new IntWritable(rating));
+            double rating = Double.parseDouble(user_move_rating[2]);
+            context.write(new Text(user_move_rating[0]), new DoubleWritable(rating));
         }
     }
 
-    public static class NullReducer extends Reducer<Text, IntWritable, Text, IntWritable>{
+    public static class NullReducer extends Reducer<Text, DoubleWritable, Text, DoubleWritable>{
         private int numOfMovies;
         @Override
         public void setup(Context context){
@@ -33,16 +34,16 @@ public class NullValueHandler {
             numOfMovies = conf.getInt("movies", 1);
         }
         @Override
-        public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException{
+        public void reduce(Text key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException{
             //key: user
             //value: rating
-            int sum=0;
+            double sum=0.0;
 
             while (values.iterator().hasNext()){
                 sum += values.iterator().next().get();
             }
-            int average = sum/numOfMovies;
-            context.write(key, new IntWritable(average));
+            double average = sum/numOfMovies;
+            context.write(key, new DoubleWritable(average));
         }
     }
 
@@ -59,7 +60,7 @@ public class NullValueHandler {
         job.setInputFormatClass(TextInputFormat.class);
         job.setOutputFormatClass(TextOutputFormat.class);
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(IntWritable.class);
+        job.setOutputValueClass(DoubleWritable.class);
 
         TextInputFormat.setInputPaths(job, new Path(args[0]));
         TextOutputFormat.setOutputPath(job, new Path(args[1]));
